+++ modules/Grades/EditReportCardGrades.php
@@ -146,8 +146,8 @@ if ( UserStudentID() )
 			}
 		}
 
-		// Unset modfunc & redirect URL.
-		RedirectURL( 'modfunc' );
+		// Unset modfunc, values & redirect URL.
+		RedirectURL( [ 'modfunc', 'values' ] );
 	}
 
 	if ( $_REQUEST['modfunc'] === 'remove' )