+++ modules/Scheduling/Requests.php
@@ -87,8 +87,8 @@ if ( $_REQUEST['modfunc'] === 'update' )
 		}
 	}
 
-	// Unset modfunc & redirect URL.
-	RedirectURL( 'modfunc' );
+	// Unset modfunc, values & redirect URL.
+	RedirectURL( [ 'modfunc', 'values' ] );
 }
 
 // Add.