+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 12.0
 - Check `$_REQUEST` after modfunc in case it is empty (no form inputs have changed), program wide
 - SQL Use DBInsert() function instead of DBInsertSQL() in AddActivity.php, Calendar.php, MarkingPeriods.php, MassAssignFees.php & MassAssignPayments.php
 - Fix SQL syntax error at or near ")" when selected MP not in current School Year in PrintSchedules.php
+- Unset `$_REQUEST` data (values, student, ...) array once modfunc done, program wide
 
 Changes in 11.8.1
 -----------------